+++ src/config.rs
@@ -289,6 +289,7 @@ impl Config {
                 axum::http::header::AUTHORIZATION,
                 axum::http::header::CONTENT_TYPE,
                 axum::http::header::ACCEPT,
+                axum::http::HeaderName::from_static("x-view-as-user"),
             ]))
     }
 }